fix: complete Image replacement, deduplicate wallet types, unify connect logic, remove next-themes
