+++ package.json
@@ -24,7 +24,6 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.475.0",
-    "next-themes": "^0.4.4",
     "react": "^18.0.0",
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^18.0.0",
