+++ packages/create-w3-kit/components/network-switcher/utils.ts
@@ -1,14 +1,5 @@
 import { Network } from './types';
-
-interface EIP1193Provider {
-  request: (args: { method: string; params?: unknown[] }) => Promise<unknown>;
-}
-
-declare global {
-  interface Window {
-    ethereum?: EIP1193Provider;
-  }
-}
+import '../../../../lib/ethereum';
 
 export async function switchNetwork(network: Network): Promise<void> {
   if (!window.ethereum) {
