+++ packages/create-w3-kit/components/token-list/TokenList.tsx
@@ -37,7 +37,7 @@ const TokenImage = memo(({ logoURI, symbol, size = "md" }: { logoURI?: string; s
   
   return (
     <div className={`${sizeClasses[size]} rounded-full overflow-hidden flex-shrink-0 bg-gray-100 dark:bg-gray-700 border border-gray-200 dark:border-gray-600 relative`}>
-      <Image
+      <img loading="lazy"
         src={logoURI}
         alt={symbol}
         width={32}
