+++ registry/w3-kit/smart-contract-scanner/smart-contract-scanner.tsx
@@ -723,7 +723,6 @@ export const SmartContractScanner: React.FC<SmartContractScannerProps> = ({
                 </div>
               </div>
             )}
-          </div>
           </CardContent>
         </>
       )}
