+++ registry/w3-kit/token-card/token-card.tsx
@@ -165,10 +165,9 @@ export const TokenCard: React.FC<TokenCardProps> = ({
           <div className="flex items-center space-x-3">
             <div className="relative flex-shrink-0">
               <div className="relative w-10 h-10 rounded-full overflow-hidden bg-secondary shadow-sm">
-                <Image
+                <img loading="lazy"
                   src={logoURI}
                   alt={token.symbol}
-                  fill
                   className="object-cover"
                 />
               </div>
@@ -472,7 +471,7 @@ export const TokenCard: React.FC<TokenCardProps> = ({
       >
         <div className="flex items-center space-x-2.5">
           <div className="relative">
-            <Image
+            <img loading="lazy"
               src={logoURI}
               alt={token.symbol}
               width={30}
@@ -664,7 +663,7 @@ export const TokenCard: React.FC<TokenCardProps> = ({
             <div className="flex items-center justify-between mb-3">
               <div className="flex items-center space-x-3">
                 <div className="relative">
-                  <Image
+                  <img loading="lazy"
                     src={logoURI}
                     alt={token.symbol}
                     width={30}
@@ -945,7 +944,7 @@ export const TokenCard: React.FC<TokenCardProps> = ({
         onClick={() => onClick?.(token)}
       >
         <div className="relative">
-          <Image
+          <img loading="lazy"
             src={logoURI}
             alt={token.symbol}
             width={18}
